moved pianokeyboard into split keyroll components
